Added Devise Configurations and User Model

+++ config/application.rb
@@ -1,13 +1,15 @@
-require_relative "boot"
+# frozen_string_literal: true
 
-require "rails/all"
+require_relative 'boot'
+
+require 'rails/all'
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
 
 module UniversityMain
-  class Application < Rails::Application
+  class Application < Rails::Application # rubocop:disable Style/Documentation
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 7.0
 
@@ -20,9 +22,11 @@ module UniversityMain
     # config.eager_load_paths << Rails.root.join("extras")
     # credential for only development env
     env_file = File.join(Rails.root, 'config', 'config_env.yml')
-    YAML.load(File.open(env_file)).each do |key, value|
-      ENV[key.to_s] = value
-    end if File.exist?(env_file) && Rails.env.upcase == 'DEVELOPMENT'
+    if File.exist?(env_file) && Rails.env.upcase == 'DEVELOPMENT'
+      YAML.load(File.open(env_file)).each do |key, value|
+        ENV[key.to_s] = value
+      end
+    end
 
     # Only loads a smaller set of middleware suitable for API only apps.
     # Middleware like session, flash, cookies can be added back manually.
